\name{fit2csv}
\Rdversion{2.0}
\alias{fit2csv}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Output a processed track to .csv files
}
\description{
Outputs a processed KF object to three .csv files : 1) the track 2) the initial and estimated parameters 3) the variance structure, max gradient and call.
}
\usage{
fit2csv(fit, name = 'fit.csv', folder = getwd())
}

\author{
Chi Hin (Tim) Lam
}

\examples{

}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{}