% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.erddap.R
\name{get.erddap}
\alias{get.erddap}
\title{Title}
\usage{
get.erddap(track, every.day = 5, folder = tempdir(),
  server = "http://coastwatch.pfeg.noaa.gov/erddap", type = 2,
  product = "ncdcOisst2Agg", variable = "sst", at = 1, depth = c(0, 0),
  yearless = F, kml.image = F, colorbar = "|||||", full.extent = F,
  name = NA, repack = TRUE, trim = TRUE, mybox = NA)
}
\arguments{
\item{track}{}

\item{every.day}{}

\item{folder}{}

\item{server}{}

\item{type}{}

\item{product}{}

\item{variable}{}

\item{at}{}

\item{depth}{}

\item{yearless}{}

\item{kml.image}{}

\item{colorbar}{}

\item{full.extent}{}

\item{name}{}

\item{repack}{}

\item{trim}{}
}
\description{
Title
}
