% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWTextract.r
\name{MWTextract}
\alias{MWTextract}
\title{Read in data from a Microwave Telemetry Inc. fish report from a PSAT.
Reads in all data from a Microwave Telemetry Inc. fish report and parses all pieces into a list. Currently, the only information not readin is the light based locations while the PSAT tag is floating at the surface.}
\usage{
MWTextract(tagID, xlsfile, delta = F, minmax = F)
}
\arguments{
\item{tagID}{Tag ID}

\item{xlsfile}{.xlsx version of MWT fish report}

\item{delta}{Logical. Should the delta limited temp/depths be flagged and removed}

\item{minmax}{Logical. This should be T if the tabs are present. Older fish reports did not have this data feature.}
}
\value{
a list with MWT PSAT header with the following:
\item{x0}{Tagging Location}
\item{day0}{Tagging date of Class 'date'} 
\item{day0b}{Tagging date of Class 'POSIXct'}
\item{xT}{pop-off location}
\item{dayT}{pop-off Date of Class 'date'}
\item{dayTb}{pop-off Date of Class 'POSIXct'}
\item{fulldates}{Alldates fish was at liberty of Class 'date'. This does not neccesarily correspond to how many days of data are returned}
\item{ SRSS }{Data frame of Sunrise and Sunset times. Columns are: Date, Sunrise and Sunset with SR and SS in minutes past midnight. Date is of Class'date' }  
\item{MWTxy}{Data Frame of Microwave Telemetry produced light based geolocations and corresponding dates. Columns are: Year, Month, Day,Latitude and Longitude (-180 to 180)}
\item{T}{Matrix of ambient temperatures. Columns correspond to time of day GMT (typically 0-23.75 hours) and Rows corrspond to days at liberty. This matrix contains both observations and any times missing observations}
\item{Z}{Matrix of depth. Columns correspond to time of day GMT (typically 0-23.75 hours) and Rows corrspond to days at liberty. This matrix contains both observations and any times missing observations}     
\item{Argos}{Data Frame of Argos data returned while tag was floating. Columns are: Date-Time (character class), Argos Location Class, Latitude and Longitude.}
}
\description{
This function uses the \code{\link{readxl}} library. The fish report MUST be converted into the newest Excel format (.xlsx). Otherwise, RStudio will crash (email Hadley WIckham to complain..).
}
\examples{
# none
}
