% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSSTlegend.r
\name{plotSSTlegend}
\alias{plotSSTlegend}

\title{plotSSTlegend}
\usage{
plotSSTlegend(cbrks = 33, cex = 1.2, font = 1)
}
\arguments{
\item{cbrks}{}

\item{cex}{}

\item{font}{}
}
\description{

plotSSTlegend
}
