% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTZhist.r
\name{plotThist}
\alias{plotThist}
\title{Title}
\usage{
plotThist(allTZ, tbrks = seq(3, 35, by = 1), xdiv = 4, xseq = 0.25, ...)
}
\arguments{
\item{allTZ}{}

\item{tbrks}{}

\item{xdiv}{}

\item{xseq}{}

\item{...}{}
}
\description{
Title
}
