% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTZhist.r
\name{plotZhist}
\alias{plotZhist}
\title{plotZhist}
\usage{
plotZhist(allTZ, interval = 10, plotmin = -200, xdiv = 4, xseq = 0.25,
  ...)
}
\arguments{
\item{allTZ}{}

\item{interval}{}

\item{plotmin}{}

\item{xdiv}{}

\item{xseq}{}

\item{...}{}
}
\description{
plotZhist
}
