% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepf.R
\name{prepf}
\alias{prepf}
\title{Generate an object for kftrack/ukfsst estimation
This function takes an object created by \code{\link{MWTextract}} and returns a data frame with 7 columns suitable for kf estimation. This also sets some limits on where in the world the fish may have gone. These limits are meant to be extremely broad and exist to filter only the most egregiously erroneous light measurements. The result is a standard format for kftrack/ukfsst analysis}
\usage{
prepf(tag, xmin = -100, xmax = 0, ymin = 10, ymax = 55, keepall = F,
  sst.depth = NULL, use.minmax = F)
}
\arguments{
\item{tag}{object returned from MWTextract}

\item{xmin}{Western longitude cutoff}

\item{xmax}{Eastern longitude cutoff}

\item{ymin}{Southern latitude cutoff}

\item{ymax}{Northern latitude cutoff}

\item{keepall}{logical. If you do not wish to use these cutoffs, keepall = T. Default is keepall = F}

\item{use.minmax}{Should min/max temperature and depth tabs be used. If False, the archival data will be used to detemine a min/max for each day. If the minmax data are present, this should be True.}
}
\value{
a dataframe
}
\description{
Generate an object for kftrack/ukfsst estimation
This function takes an object created by \code{\link{MWTextract}} and returns a data frame with 7 columns suitable for kf estimation. This also sets some limits on where in the world the fish may have gone. These limits are meant to be extremely broad and exist to filter only the most egregiously erroneous light measurements. The result is a standard format for kftrack/ukfsst analysis
}
\examples{
tagdata = MWTextract(tagID = '12345',  xlsfile = '12345.xlsx', use.minmax = T)
xdata = prepf(tagdata, xmin = 100, xmax = 0, ymin = 10, ymax = 55, keepall = F)

}
