% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite.kml.R
\name{satellite.kml}
\alias{satellite.kml}
\title{Title}
\usage{
satellite.kml(track, folder = getwd(), local = F, product = "TBAssta",
  interval = 5,
  server = "http://coastwatch.pfeg.noaa.gov/coastwatch/CWBrowserWW360.jsp?get=gridData&dataSet=")
}
\arguments{
\item{track}{}

\item{folder}{}

\item{local}{}

\item{product}{}

\item{interval}{}

\item{server}{}
}
\description{
Title
}
